+++ tests/e2e/cases/process_stdio_readable_writable/in.ts
@@ -1,3 +1,5 @@
 console.log("stdin readable:", process.stdin.readable);
+console.log("stdout readable:", (process.stdout as any).readable);
+console.log("stderr readable:", (process.stderr as any).readable);
 console.log("stdout writable:", process.stdout.writable);
 console.log("stderr writable:", process.stderr.writable);